fix: Fix .gitconfig "resource busy" issue
It looks like git overwrites the .gitconfig file when it changes something. This is not possible however because the file is mounted to the host file system, so it cannot be replaced, only the content can be changed.
The fix is to mount it somewhere else (read-only) and then copy it to the desired place at container start.
Because everything in the user's home directory is now a copy and not directly mounted anymore, this also has the side effect that the user can now freely experiment with the .gitconfig in the container.

Also:
- FIx syntax to mount read-only in docker-compose.
  False: `readonly`
  Correct: `ro`